#2216: Update fmt to 10.1

+++ src/vt/utils/compress/decompressor.impl.h
@@ -136,7 +136,7 @@ std::size_t Decompressor<Readable>::read(
       // we have hit an unknown error, print the code and corresponding message!
       auto error_code = BrotliDecoderGetErrorCode(dec_);
       auto error_str = fmt::format(
-        "code={}, msg={}\n", error_code, BrotliDecoderErrorString(error_code)
+        "code={}, msg={}\n", static_cast<int32_t>(error_code), BrotliDecoderErrorString(error_code)
       );
       vtAbort(error_str);
       break;
